support customer-specific l2

+++ tests/conftest.py
@@ -11,11 +11,12 @@ import traceback
 from pathlib import Path
 from subprocess import check_call, getoutput
 from typing import List
-from unittest.mock import patch
+from unittest.mock import Mock, patch
 
 import mock
 import pytest
 import structlog
+from structlog.typing import WrappedLogger
 
 import fc.qemu.agent
 import fc.qemu.hazmat.qemu
@@ -23,7 +24,8 @@ import fc.qemu.logging
 from fc.qemu.agent import Agent
 from fc.qemu.hazmat import libceph
 from fc.qemu.hazmat.ceph import Ceph, RootSpec, VolumeSpecification
-from fc.qemu.util import GiB
+from fc.qemu.hazmat.iproute2 import TunTap
+from fc.qemu.util import GiB, cmd
 
 ########################################################
 # ceph fixtures
@@ -546,7 +548,7 @@ def named_vm(name, request, clean_environment, monkeypatch, tmpdir):
     def cleanup():
         if p := vm.qemu.proc():
             parent = p.parent()
-            if "supervised-qemu-wrapped" in parent.cmdline()[1]:
+            if parent and "supervised-qemu-wrapped" in parent.cmdline()[1]:
                 parent.terminate()
             p.terminate()
 
@@ -696,28 +698,31 @@ def setup_structlog():
     # set to True to temporarily get detailed tracebacks
     log_exceptions = True
 
-    def test_logger(logger, method_name, event):
+    def test_logger(
+        logger: WrappedLogger,
+        method_name: str,
+        event: fc.qemu.logging.EventDict,
+    ):
         stack = event.pop("stack", None)
         exc = event.pop("exception", None)
         event_name = event.pop("event", "")
         event_prefix = os.path.basename(event_name) if event_name else " "
-        result = []
+
+        result: str
+        # In reality we log via `output` but in the tests we log via `output_line` to help
+        # with diagnosis on partially completed/hanging calls.
+        event.pop("output", None)
         if "output_line" in event:
             result = fc.qemu.logging.prefix(
-                event_prefix, event["output_line"].strip()
+                event_prefix, event.pop("output_line").rstrip()
             )
         else:
-            output = event.pop("output", None)
-
-            result = []
+            result_: list[str] = []
             if event_name:
-                result.append(event_name)
+                result_.append(event_name)
             for key in sorted(event):
-                result.append("{}={}".format(key, str(event[key]).strip()))
-            result = " ".join(result)
-
-            if output:
-                result += fc.qemu.logging.prefix(event_prefix, output)
+                result_.append("{}={}".format(key, str(event[key]).strip()))
+            result = " ".join(result_)
 
         # Ensure we get something to read on stdout in case we have errors.
         reltime = time.time() - util.test_log_start
@@ -878,3 +883,17 @@ def pytest_assertrepr_compare(op, left, right):
         return left.compare(right).diff
     elif right.__class__.__name__ == "Ellipsis":
         return right.compare(left).diff
+
+
+@pytest.fixture(autouse=True)
+def cleanup_tap_devices():
+    for tuntap in TunTap.list(Mock()):
+        cmd(f"ip l delete dev {tuntap.ifname}", Mock())
+    yield
+    try:
+        for tuntap in TunTap.list(Mock()):
+            cmd(f"ip l delete dev {tuntap.ifname}", Mock())
+    except Exception:
+        # Optimistic here: some tests patch out `cmd` and this may
+        # not have been cleaned up yet (e.g. test_rbd_unexpected_exception_does_not_cause_image_not_found)
+        pass
